Add built-in compilation unit with Any type decl
An Any type declaration (AnyTypeDecl) is needed in the AST, so that uses
of the type can be automatically resolved without adding code which
treats uses of the Any type as special cases, as it is currently done in
the `commonSupertype` branch.

However, there is still a bug in
org.abs_models.frontend.parser.Main::getBuiltinUnit which causes tests
to fail.